# config/swagger.json
{
    "openapi": "3.0.0",
...
    "servers": [
        {
            "url": "{protocol}://localhost:{port}",
            "description": "Local development server",
            "variables": {
                "protocol": {
                    "enum": ["http"],
                    "default": "http"
                },
                "port": {
                    "enum": ["3000"],
                    "default": "3000" 
                }
            }
        }
    ],
    "paths": {
...
        "/auth/register": {
            "get": {
                "summary": "Returns register page",
                "responses": {
                    "200": {
                        "description": "OK: Returns EJS register template",
                        "content": {
                            "html/text": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates a user",
                "description": "Redirect to /auth/login if user was created or try again",
                "requestBody": {
                    "description": "Credentials from registration form fields: username, password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentials"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentials"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED: User was created. Redirect to login page"
                    },
                    "400": {
                        "description": "BAD REQUEST: Username/password wasn't provided"
                    },
                    "409": {
                        "description": "CONFLICT: User already exists"
                    }
                }
            }
        },
        "/auth/login": {
            "get": {
                "summary": "Returns login page",
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK: Returns EJS login template",
                        "content": {
                            "html/text": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Login user",
                "description": "Login and redirect to profile page or try again",
                "requestBody": {
                    "description": "Credentials from login form fields: username, password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentials"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCredentials"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK: LogIn and redirect to user profile"
                    },
                    "401": {
                        "description": "UNAUTHORIZED: Incorrect username/password"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserCredentials": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": ["username", "password"]
            }
        }
    }   
}